% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_methods.R
\name{ab-methods}
\alias{ab-methods}
\alias{summary.ab}
\alias{print.summary.ab}
\alias{print.ab}
\title{Methods for ab objects}
\usage{
\method{summary}{ab}(object, digits = 3, ...)

\method{print}{summary.ab}(x, ...)

\method{print}{ab}(x, ...)
}
\arguments{
\item{object, x}{object of class \code{ab} as returned from
\code{\link{ab_test}}.}

\item{digits}{number of digits to print for the summary.}

\item{...}{further arguments, currently ignored.}
}
\value{
The \code{summary} methods returns the \code{ab} object that is
  guaranteed to contain posterior samples (i.e., it adds posterior samples if
  they were not included already). Additionally, the list output contains the
  number of digits to print for the posterior summary.

  The \code{print} methods simply print and return nothing.
}
\description{
Methods defined for objects returned from the \code{\link{ab_test}} function.
}
