% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{calc_beta}
\alias{calc_beta}
\title{Find Beta distribution shape parameters}
\usage{
calc_beta(x1 = 0.05, p1 = 0.025, x2 = 0.95, p2 = 0.975)
}
\arguments{
\item{x1}{First proportion}

\item{p1}{Quantile value of first proportion}

\item{x2}{Second proportion}

\item{p2}{Quantile value of second proportion}
}
\value{
A list with two elements representing the two shape parameters of
the Beta distribution.
}
\description{
Given two proportions and their quantile values, find the Beta shape
parameters that represent the distribution.
}
\examples{
# Find the Beta distribution where the 2.5 percentile is 0.08 and the 97.5
# percentile is 0.42.
calc_beta(x1 = 0.08, p1 = 0.025, x2 = 0.42, p2 = 0.975)
}
