% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clt_transform}
\alias{clt_transform}
\title{Central Limit Theorem transformation}
\usage{
clt_transform(x1)
}
\arguments{
\item{x}{data to be transformed}
}
\value{
transformed vector
}
\description{
According to CLT we sample multiple times to produce a vector of sample means
}
