% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_comparison.R
\name{hist_comparison}
\alias{hist_comparison}
\title{Compare histograms between different groups}
\usage{
hist_comparison()
}
\value{
Side by side histograms
}
\description{
This function plots side by side the histograms of data
}
