% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{logit_beta}
\alias{logit_beta}
\title{Beta distribution logit scale prior}
\usage{
logit_beta(shape1, shape2, n = 10000)
}
\arguments{
\item{shape1}{First shape parameter for Beta distribution}

\item{shape2}{Second shape parameter for Beta distribution}

\item{n}{The number of draws to make from the Beta distribution}
}
\value{
A list with two elements representing the location and scale
parameters of the Normal distribution.
}
\description{
For a given Beta distribution find the mean and standard deviation of a
Normal distribution that can be used as a prior for a GLM.
}
\examples{
logit_beta(5, 17, n = 1e4)
}
