% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{t_test_two_sample}
\alias{t_test_two_sample}
\title{Two Sample t-test}
\usage{
t_test_two_sample(x1, x2)
}
\arguments{
\item{x1, x2}{samples}
}
\value{
p-value of test along with a histogram
}
\description{
Compare the means of two samples
}
