% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{t_test_one_sample}
\alias{t_test_one_sample}
\title{One Sample t-test}
\usage{
t_test_one_sample(x, m0, clt = F)
}
\arguments{
\item{x}{sample}

\item{m0}{value to be tested against}

\item{clt}{applies or not clt_transform function}
}
\value{
p-value of test along with a histogram
}
\description{
Compare the mean of the sample with a value
}
