% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{t_test_two_sample}
\alias{t_test_two_sample}
\title{Two Sample t-test}
\usage{
t_test_two_sample(x1, x2, clt = F)
}
\arguments{
\item{x1, x2}{samples to be compared}

\item{clt}{applies or not clt_transform function}
}
\value{
p-value of test along with a histogram
}
\description{
Compare the means of two samples
}
